+++ spacepackets/cfdp/pdu/ack.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 import enum
-from typing import cast
 
 from spacepackets.cfdp.pdu import PduHeader
 from spacepackets.cfdp.pdu.file_directive import (
